move get|setLineBgColor to Shape as it is only implemented there
